+++ History.md
@@ -1,5 +1,11 @@
 
-0.0.1 / 2010-08-02
+0.0.1 / 2010-08-26
 ------------------
 
-* jquery-conflict-sandbox initial release
+* jquery-sandbox initial release
+
+
+0.0.2 / 2010-08-26
+------------------
+
+* jquery-sandbox API change